app/requirements.txt
asgiref==3.3.1
Django==3.1.7
djangorestframework==3.11.1
pytz==2021.1
sqlparse==0.4.1
djoser==2.0.5
psycopg2-binary==2.8.6
whitenoise==5.2.0
dj-database-url==0.5.0
django-cors-headers==3.7.0